ProjectFiles/pyproject.toml
[project]
name = "WindCore"
version = "0.1.0"
description = "Weather-Based Prediction of Wind Turbine Energy Output"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "joblib>=1.5.3",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "iprogress>=0.4",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
]
